+++ src/main/java/com/viraj/sample/repository/AvionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AvionRepository extends CrudRepository<Avion,Long> {
+public interface AvionRepository extends CrudRepository<Avion,Integer> {
     
 }
